Rename unprocessed_transaction_storage to vote_storage
